+++ CHANGELOG.md
@@ -1,9 +1,23 @@
 # CHANGELOG
 
-## Version 1.0.0
+All notable changes to this project will be documented in this file.
 
-Initial release.
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Version 1.0.1
+## [Unreleased]
+
+### Added
+
+- Added the ability to plot multiple fitting histories returned by the [`fit`](https://keras.io/api/models/model_training_apis/#fit-method) function on a single figure.
+
+### Changed
+
+- The `history` parameter of the `lcurves_by_history` function can now accept a list of dictionaries with multiple fitting histories of keras models.
+
+## [1.0.1] - 2025-01-23
 
 The default value for the `figsize` parameter of the `lcurves_by_history` and `lcurves_by_MLP_estimator` functions has been changed to `None`.
+
+## [1.0.0] - 2024-01-22
+
+Initial release.